edit user API components

+++ src/main/kotlin/com/project/lunchBooking/repo/UserRepository.kt
@@ -3,5 +3,6 @@ package com.project.lunchBooking.repo
 import com.project.lunchBooking.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository: JpaRepository<User, Long> {
+interface UserRepository: JpaRepository<User, Int> {
+    
 }
